requirements.txt
django>=5.2,<6.0
djangorestframework>=3.17,<4.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
anthropic>=0.96,<1.0
requests>=2.33,<3.0
gunicorn>=23.0,<24.0
whitenoise>=6.8,<7.0
dj-database-url>=2.3,<3.0
